requirements.txt
pillow>=10.0.0
pycryptodome>=3.19.0
opencv-python>=4.12.0
